wasm fallout
